+++ examples/erc20-balance-history/entities.ts
@@ -1,4 +1,4 @@
-import { createEntity } from 'https://deno.land/x/robo_arkiver@v0.4.14/mod.ts'
+import { createEntity } from 'https://deno.land/x/robo_arkiver@v0.4.15/mod.ts'
 
 // @note: "Index: true" enhances graphql queries