+++ examples/erc20-balance-history/manifest.ts
@@ -1,4 +1,4 @@
-import { Manifest } from 'https://deno.land/x/robo_arkiver@v0.4.14/mod.ts'
+import { Manifest } from 'https://deno.land/x/robo_arkiver@v0.4.15/mod.ts'
 import erc20 from './erc20.ts'
 import { Entities } from './entities.ts'
 import { onTransfer } from './handlers.ts'
@@ -7,8 +7,8 @@ const manifest = new Manifest('frax-balances')
 
 manifest
   .addEntities(Entities)
-  .chain('mainnet', { blockRange: 500n })
-  .contract(erc20)
+  .addChain('mainnet', { blockRange: 500n })
+  .addContract(erc20)
   .addSources({ '0x853d955aCEf822Db058eb8505911ED77F175b99e': 11465581n })
   .addEventHandlers({ 'Transfer': onTransfer })